utils: Refactor internal utils
Moved the file pattern recognition function and others having global meaning to the API.

Conversion function `to_lowercase()` was separated to the new internal `convert.hpp` file.
